* enhancement_269_refactoring_for_adds_on: (93 commits)
  Some changes to prevent inadvertantly saving temporary to session user settings for anonymous user. Mostly in the user model methods.
  Getting rid of the ability to change the role or take action on the anonymous account.
  Wrapping up test refactoring for now and made some changes to have dynamic reader method be what is expected in most cases. Extended multiple ftype text fields return an array of arrays with single value (used internally as the data structure for a number of ftype field processing) when what is expected would be to  return an array of string values. Living with this for now (other higher priorities), but keeping a failed test in place to nag about it.
  reverting debugging that was added by mistake
  fixing bug in external search sources (updated from its repo)
  merging
  Refining expected data structure after testing against master for reference.
  Added helper method.
  Some more slight DRYing/refactoring and added accessor tests for text and (failing, as expected) ftype topic_type.
  Refactored out a good bit of duplication that was bugging me. Should make these tests easier to read and maintain. Could do more, but this seems ok for now.
  Cleaning up test. Spoke to soon as it pointing to where things are failing for ftype = topic_type. More tests coming...
  Got rid of some duplication in existing tests with new helper and added new ftype multple test. This is failing and may point at our issue that we are investigating.
  Fixing tests for choices and adding a new ftype = topic_type test.
  Added extended field topic type validity test.
  remove debugger and more explicit check on fields
  Updated extended_content for acessing choices
  Adding check for url being not nil.
  Fixed last set of tests.  Bit of a hack (clearing zebra before test) and unfortunately slows test performance. Something to clean up other ways later.
  Missed committing this hardcoding replacement before.
  Replaced failing hardcoded text with locale lookup.
  ...

+++ test/integration/extended_content_test.rb
@@ -126,7 +126,7 @@ class ExtendedContentTest < ActionController::IntegrationTest
         visit "/site/content_types/edit/#{content_type_ids[class_name]}"
 
         verb = options[:extended_field_value_required] ? "required" : "add"
-        check "extended_field_#{ExtendedField.last.id.to_s}_#{verb}_checkbox"
+        check "extended_field_#{ExtendedField.last.to_param.to_s}_#{verb}_checkbox"
 
         click_button "Add to Content Type"
 
@@ -139,7 +139,7 @@ class ExtendedContentTest < ActionController::IntegrationTest
       end
 
       teardown do
-        ContentTypeToFieldMapping.last.destroy
+        ContentTypeToFieldMapping.last.destroy if ContentTypeToFieldMapping.last
         @extended_field.destroy
       end
 
@@ -234,13 +234,8 @@ class ExtendedContentTest < ActionController::IntegrationTest
       @topic.attributes = {
         :title => 'Choice Linking Test',
         :topic_type_id => @topic_type.id,
-        :basket_id => @@site_basket.id,
-        :extended_content_values => {
-          "home_town" => {
-            "1" => { "preset" => "", "custom" => "Somewhere way out there" }
-          }
-        }
-      }
+        :basket_id => @@site_basket.id }
+      @topic.home_town = "Somewhere way out there"
       @topic.save
       @topic.creator = @james
     end
@@ -345,7 +340,7 @@ class ExtendedContentTest < ActionController::IntegrationTest
       click_button "Create"
 
       verb = options[:extended_field_value_required] ? "required" : "add"
-      check "extended_field_#{ExtendedField.last.id.to_s}_#{verb}_checkbox"
+      check "extended_field_#{ExtendedField.last.to_param.to_s}_#{verb}_checkbox"
       click_button "Add to Topic Type"
 
       body_should_contain "#{options[:extended_field_label]}"
